Q42.py
This file contains a python code that includes the function found in Q41.py which is in this repository. This function will find how many times a specific substring occurs inside a string and then if this number is greater than or equal to 4 it will print a certain string and if it is less than 4 it will print something different.